add jiemian
add jiemian

+++ del.c
@@ -10,7 +10,9 @@ int cgiMain()
 	fprintf(cgiOut, "Content-type:text/html;charset=utf-8\n\n");
 
 	char stuId[32] = "\0";
+	char sstatus[4] ="\0";
 	int status = 0;
+
 
 
 	status = cgiFormString("stuId",  stuId, 32);
@@ -19,7 +21,12 @@ int cgiMain()
 		fprintf(cgiOut, "get stuId error!\n");
 		return 1;
 	}
-
+	status = cgiFormString("sstatus",  sstatus, 4);
+	if (status != cgiFormSuccess)
+	{
+		fprintf(cgiOut, "get sstatus error!\n");
+		return 1;
+	}
 
 	int ret;
 	char sql[128] = "\0";
@@ -43,7 +50,7 @@ int cgiMain()
 	}
 mysql_options(db,MYSQL_SET_CHARSET_NAME,"utf8");
 
-
+if (sstatus[0]=='1') {
 	sprintf(sql, "delete from Student where stuID = %d", atoi(stuId));
 	if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
 	{
@@ -51,6 +58,18 @@ mysql_options(db,MYSQL_SET_CHARSET_NAME,"utf8");
 		mysql_close(db);
 		return ;
 	}
+
+}
+else{
+	sprintf(sql, "update Student set sstatus= %d where stuID = %d ", atoi(sstatus),atoi(stuId));
+	if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
+	{
+		fprintf(cgiOut,"mysql_real_query fail:%s\n", mysql_error(db));
+		mysql_close(db);
+		return ;
+	}
+}
+
 
 
 	fprintf(cgiOut, "delete stu ok!\n");
